+++ lib/gandi_v5/live_dns/domain.rb
@@ -176,6 +176,20 @@ class GandiV5
         data['message']
       end
 
+      # Get the domain's zone from Gandi.
+      # @return [GandiV5::LiveDNS::Zone]
+      # @raise [GandiV5::Error::GandiError] if Gandi returns an error.
+      def fetch_zone
+        GandiV5::LiveDNS::Zone.fetch zone_uuid
+      end
+
+      # The domain's zone (fetching from Gandi if required).
+      # @return [GandiV5::LiveDNS::Zone]
+      # @raise [GandiV5::Error::GandiError] if Gandi returns an error.
+      def zone
+        @zone ||= fetch_zone
+      end
+
       # List the domains.
       # @return [Array<GandiV5::LiveDNS::Domain>]
       # @raise [GandiV5::Error::GandiError] if Gandi returns an error.