+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 * Enhancements to GandiV5::Email::Slot
   * .create now returns created slot
   * #delete now checks for inactiveness and refundableness
+* Enhancements to GandiV5::LiveDNS::Domain
+  * Add #zone and #fetch_zone
 * Enhancements to GandiV5::LiveDNS::Zone::Snapshot
   * Add .list
   * Add .fetch